[MOVED] About fragmet to `about` package
